ddr-ai/requirements.txt
fastapi
uvicorn
streamlit
python-multipart
pdfplumber
pypdf
pytesseract
Pillow
pydantic
python-docx
python-dotenv
requests
